requirements-train.txt
torch>=2.5.0
torchvision>=0.20.0
opencv-python>=4.10.0
Pillow>=10.4.0
numpy>=2.1.0
tqdm>=4.66.0
matplotlib>=3.9.0
